+++ packages/drawer/src/DrawerTabs/DrawerTabs.tsx
@@ -10,6 +10,12 @@ import { useDrawerContext } from '../DrawerContext';
 import { getDrawerTabsStyles } from './DrawerTabs.styles';
 import { DrawerTabsProps } from './DrawerTabs.types';
 
+/**
+ * @deprecated
+ *
+ * This is deprecated and is no longer available for use. Stay tuned for an updated way to leverage information architecture within a `Drawer` component.
+ */
+
 export const DrawerTabs = ({
   children,
   className,