+++ packages/pipeline/src/TooltipText/TooltipText.types.ts
@@ -1,0 +1,3 @@
+export interface TooltipTextProps {
+  hiddenStages: Array<string | null>;
+}