+++ chat/lg-markdown/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @lg-chat/lg-markdown
 
+## 3.0.10
+
+### Patch Changes
+
+- Updated dependencies [16dda633f]
+  - @leafygreen-ui/leafygreen-provider@4.0.5
+  - @leafygreen-ui/code@17.1.2
+  - @leafygreen-ui/typography@20.1.5
+
 ## 3.0.9
 
 ### Patch Changes