+++ chat/message-feedback/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @lg-chat/message-feedback
 
+## 4.0.4
+
+### Patch Changes
+
+- Updated dependencies [b75c9b896]
+- Updated dependencies [16dda633f]
+  - @leafygreen-ui/icon-button@16.0.7
+  - @leafygreen-ui/leafygreen-provider@4.0.5
+  - @leafygreen-ui/button@23.1.3
+  - @leafygreen-ui/popover@13.0.7
+  - @leafygreen-ui/text-area@10.0.8
+  - @leafygreen-ui/typography@20.1.5
+
 ## 4.0.3
 
 ### Patch Changes