+++ packages/expandable-card/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @leafygreen-ui/expandable-card
 
+## 4.0.8
+
+### Patch Changes
+
+- Updated dependencies [b75c9b896]
+- Updated dependencies [16dda633f]
+  - @leafygreen-ui/icon-button@16.0.7
+  - @leafygreen-ui/leafygreen-provider@4.0.5
+  - @leafygreen-ui/card@12.0.6
+  - @leafygreen-ui/typography@20.1.5
+
 ## 4.0.7
 
 ### Patch Changes