+++ packages/popover/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @leafygreen-ui/popover
 
+## 13.0.7
+
+### Patch Changes
+
+- Updated dependencies [16dda633f]
+  - @leafygreen-ui/leafygreen-provider@4.0.5
+
 ## 13.0.6
 
 ### Patch Changes